Allow explicitly specifying the project and app
This avoids using Mix.Project.get(), which doesn't always work (such as
in Livebook).

But:

As far as I can tell, the app is used to put the output binary
somewhere sensible. If you're doing this in Livebook, there really isn't
anywhere, so I used "app: :zigler". This is bad.

If that's the only place that it needs to know the app, we might be able
to work around that.